feat(run): reuse function idtoken sample (#1637)
Aligning with https://github.com/googleapis/google-auth-library-nodejs/pull/1019, this adjusts the Cloud Functions idtoken sample to be used for both Cloud Run and Cloud Functions. The main downside is this files off the URL pattern fingerprints from the sample.

Note when this is incorporated into Cloud Run documentation the run/authentication sample will be deleted.