fix(run/events-storage): e2e test checking for invalid criteria (#1676)
The events-storage test copied from events-pubsub, but where pubsub has validation in the request handler, storage does not. Changing the e2e test to check for OK status instead.

Note these are very shallow "healthcheck" grade tests, we want them to go deeper but this was a path to get tests in place for all the samples quickly.

Fixes #1647.