feat: Add snippets for Spanner DML with returning clause (#2713)
Samples are provided for INSERT, DELETE, and UPDATE in both GoogleSQL
and PostgreSQL dialects. To provide a more compelling example for the
INSERT case, a generated column has been added in the "create_database"
example so that the generated value can be returned in the INSERT
examples. This changed the number of mutations generated for inserts to
the Artists table, and the commitStats test case was updated as a
result.